Removed trivial logging test.
